Wrap powf() to fix compatibility with pre-GLIB 2.27
powf(), from math.h/libm.so.6, was changed in GLIB 2.27. Compiling against glib 2.27 or newer will result in incompatibilities with any system that has an older version of glib, such as CentOS/RHEL 7 (glib 2.17). This results in the library being unable to be loaded at runtime, with an Unsatisfied link error stating that `/lib64/libm.so.6: version `GLIBC_2.27' not found (required by <snip>/libjavaHeclib.so)`

Wrap the function, and call the glib 2.2.5 version of the function, which is present in glibc 2.2.5 and newer, restoring minimum compatibility back to 2.17 and newer versions.

Ubuntu 18.04 and CentOS 8 have glib 2.27 and 2.28 respectively, so this change fixes Ubuntu 16.04 and CentOS 7 when built on Ubuntu 18.04+ or CentOS 8+. Tested on a CentOS 7 system with the default glib 2.17.